requirements.txt
Pillow==9.5.0
fpdf2==2.7.4
python-dateutil==2.8.2
tkcalendar
num2words